Autoclaim-main/autoclaim_project/server/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
passlib[bcrypt]
python-multipart
python-dotenv
easyocr
pillow
geopy
ultralytics
torch
torchvision
huggingface-hub
groq
